runtime: Fix compile with newer Clang versions
Credits to mustermaxmuller

+++ driver/runtime/rs_allocation.c
@@ -350,9 +350,9 @@ ELEMENT_AT(double3)
 ELEMENT_AT(double4)
 
 typedef unsigned long long ull;
-typedef unsigned long long ull2 __attribute__((ext_vector_type(2)));
-typedef unsigned long long ull3 __attribute__((ext_vector_type(3)));
-typedef unsigned long long ull4 __attribute__((ext_vector_type(4)));
+typedef unsigned long ull2 __attribute__((ext_vector_type(2)));
+typedef unsigned long ull3 __attribute__((ext_vector_type(3)));
+typedef unsigned long ull4 __attribute__((ext_vector_type(4)));
 
 #ifndef RS_DEBUG_RUNTIME
 SET_ELEMENT_AT_TYPE_IMPL(ull, ulong)
